工人列表 订单列表改进

+++ Hubanghu/Hubanghu/TwoViewController/HbhMakeAppointMentViewController.h
@@ -10,5 +10,5 @@
 
 @interface HbhMakeAppointMentViewController : BaseViewController
 
-- (instancetype)initWithWorkerId:(double)aWorkerId andWorkerName:(NSString *)aWorkerName;
+- (instancetype)initWithWorkerId:(int)aWorkerId andWorkerName:(NSString *)aWorkerName;
 @end
